+++ shared/src/commonMain/kotlin/com/hub/media/features/media/data/MediaWithDetails.kt
@@ -15,5 +15,11 @@ public sealed class MediaWithDetails {
         val details: BookDetailsEntity?,
     ) : MediaWithDetails()
 
-    // Planned: Movie, TVShow
+    public data class Movie(
+        override val item: MediaItemEntity,
+    ) : MediaWithDetails()
+
+    public data class TVShow(
+        override val item: MediaItemEntity,
+    ) : MediaWithDetails()
 }